backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic[email]>=2.7.1
pydantic-settings>=2.7.0

# Supabase
supabase==2.12.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# Image Generation
Pillow==11.1.0
httpx==0.28.1
rembg==2.0.57
onnxruntime==1.19.2

# OpenAI
openai==1.59.7

# Instagram
instagrapi==2.1.2
PySocks==1.7.1

# HTTP
aiohttp==3.9.3

# Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Document parsing (brand unpacking)
pdfplumber>=0.10.0
python-docx>=1.1.0

# Utils
cryptography==44.0.0
python-multipart==0.0.18
aiofiles==24.1.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
